Add support for Date
My editor normalized style by replacing spaces to tabs, so a few lines
appear as modified but they are not, really.

Hope this does not cause any issues.

TODO: Set the proper StyleIndex for dates, I couldn't figure that out.
But the generated cell value is correct, if you manually set the format
to Date in Excel, it shows the proper date.